backend/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings

# GraphRAG and LLM
llama-index-core
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface

# Document processing
PyMuPDF
python-docx
pillow
pytesseract

# Audio processing
openai-whisper
torch
torchaudio

# Image embeddings
transformers
sentence-transformers

# Vector search and reranking
faiss-cpu
rank-bm25

# Knowledge Graph (Memgraph)
gqlalchemy>=1.4.0
spacy>=3.0.0

# Utilities
numpy
requests
aiofiles

# PostgreSQL and authentication
psycopg2-binary
asyncpg
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
